chore(models-folder): move model to enabe Regle tracking - rename eslint config file - move models folder to src folder [Maintains #168094076]
