+++ .hound.yml
@@ -1,6 +1,6 @@
-jshint: 
+jshint:
   enabled: false
 eslint:
   enabled: true
-  config_file: .eslintrc.json
+  config_file: .eslintrc
   ignore_file: .eslintignore
