+++ src/models/index.js
@@ -1,8 +1,8 @@
 import { readdirSync } from 'fs';
 import { basename as _basename, join } from 'path';
 import Sequelize from 'sequelize';
-import configFile from '../config/config.js';
-import '../../config/env';
+import configFile from '../database/config/config';
+import '../config/env';
 
 const basename = _basename(__filename);
 const env = process.env.NODE_ENV || 'development';
@@ -17,15 +17,13 @@ if (config.use_env_variable) {
 }
 
 readdirSync(__dirname)
-  .filter(file => {
-    return (file.indexOf('.') !== 0) && (file !== basename) && (file.slice(-3) === '.js');
-  })
-  .forEach(file => {
+  .filter((file) => (file.indexOf('.') !== 0) && (file !== basename) && (file.slice(-3) === '.js'))
+  .forEach((file) => {
     const model = sequelize.import(join(__dirname, file));
     db[model.name] = model;
   });
 
-Object.keys(db).forEach(modelName => {
+Object.keys(db).forEach((modelName) => {
   if (db[modelName].associate) {
     db[modelName].associate(db);
   }
