+++ src/test/middlewares/validator.spec.js
@@ -1,5 +1,7 @@
 import validate from '../../middlewares/validator';
-import { mockRequest, mockResponse, sinon, expect } from '../testHelpers/config';
+import {
+  mockRequest, mockResponse, sinon, expect
+} from '../testHelpers/config';
 
 describe('Validate middleware', () => {
   it('should call the next function if no schema is pass as arguement', () => {
@@ -7,7 +9,7 @@ describe('Validate middleware', () => {
     const req = mockRequest({
       body: {}
     });
-    const next = sinon.fake()
+    const next = sinon.fake();
     validate()(req, res, next);
     expect(next.calledOnce).to.be.true;
   });
